Split mpi.py into a self-contained module with MpiPartition, and a separate mpi_solve.py
