+++ tests/simsopt/core/test_least_squares_solve.py
@@ -4,7 +4,8 @@ from simsopt.core.functions import Identity, Rosenbrock
 from simsopt.core.optimizable import Target
 from simsopt.core.least_squares_problem import LeastSquaresProblem, LeastSquaresTerm
 from simsopt.core.serial_solve import least_squares_serial_solve
-from simsopt.core.mpi import MpiPartition, least_squares_mpi_solve
+from simsopt.core.mpi import MpiPartition
+from simsopt.core.mpi_solve import least_squares_mpi_solve
 
 def mpi_solve_1group(prob, **kwargs):
     least_squares_mpi_solve(prob, MpiPartition(ngroups=1), **kwargs)
